pyproject.toml
[project]
name = "lab3-metadata"
version = "0.1.0"
description = "Лабораторная с airflow"
authors = [{ name = "INovikov Home", email = "nov.ilucha@yandex.ru" }]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
  "apache-airflow[standard,async,celery,postgres] (>=3.1.0,<4.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
